split Paxifax into two files, with Paxifax_Aux containing supporting classes for the main types of packet processors;
