+++ Pax.csproj
@@ -26,6 +26,7 @@
     <Compile Include="AssemblyInfo.cs" />
     <Compile Include="PaxConfig.cs" />
     <Compile Include="ForwardingDecision.cs" />
+    <Compile Include="Paxifax_Aux.cs" />
     <Compile Include="Paxifax.cs" />
     <Compile Include="Pax.cs" />
     <None Include="lib/SharpPcap.dll.config">
